requirements.txt
flask==3.1.1
flask-socketio==5.5.1
python-socketio==5.13.0
python-engineio==4.12.2
eventlet==0.35.2

# RAG and AI dependencies
openai==1.70.0
pgvector==0.2.5
PyPDF2==3.0.1
python-docx==1.1.0
sqlalchemy==2.0.23
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-mail==0.9.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9